Moving funcitons out of Model into Importer or utilities.cpp
