+++ src/importer.cpp
@@ -16,7 +16,6 @@ limitations under the License.
 
 #include "libcellml/importer.h"
 
-
 #include <algorithm>
 #include <cmath>
 #include <libxml/uri.h>
@@ -34,7 +33,6 @@ limitations under the License.
 #include "xmldoc.h"
 #include "xmlutils.h"
 
-
 namespace libcellml {
 
 /**
@@ -46,7 +44,6 @@ namespace libcellml {
  */
 struct Importer::ImporterImpl
 {
-
 };
 
 Importer::Importer()
@@ -62,6 +59,138 @@ ImporterPtr Importer::create() noexcept
 Importer::~Importer()
 {
     delete mPimpl;
+}
+
+void flattenComponent(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t index)
+{
+    if (component->isImport()) {
+        auto model = owningModel(component);
+        auto importSource = component->importSource();
+        auto importModel = importSource->model();
+        auto importedComponent = importModel->component(component->importReference());
+
+        // Determine names of components already in use.
+        NameList compNames = componentNames(model);
+
+        // Determine the stack for the destination component.
+        IndexStack destinationComponentBaseIndexStack = reverseEngineerIndexStack(component);
+
+        // Determine the stack for the source component.
+        IndexStack importedComponentBaseIndexStack = reverseEngineerIndexStack(importedComponent);
+
+        // Generate equivalence map for the source component.
+        EquivalenceMap map;
+        recordVariableEquivalences(importedComponent, map, importedComponentBaseIndexStack);
+        generateEquivalenceMap(importedComponent, map, importedComponentBaseIndexStack);
+
+        // Rebase the generated equivalence map from the source component to the destination component.
+        auto rebasedMap = rebaseEquivalenceMap(map, importedComponentBaseIndexStack, destinationComponentBaseIndexStack);
+
+        // Take a copy of the imported component which will be used to replace the import defined in this model.
+        auto importedComponentCopy = importedComponent->clone();
+        importedComponentCopy->setName(component->name());
+        for (size_t i = 0; i < component->componentCount(); ++i) {
+            importedComponentCopy->addComponent(component->component(i));
+        }
+
+        // Get list of required units from component's variables.
+        std::vector<UnitsPtr> requiredUnits = unitsUsed(importModel, importedComponentCopy);
+
+        // Add all required units to a model so referenced units can be resolved.
+        auto requiredUnitsModel = Model::create();
+        for (const auto &units : requiredUnits) {
+            requiredUnitsModel->addUnits(units);
+        }
+
+        // Make a map of component name to component pointer.
+        ComponentNameMap newComponentNames = createComponentNamesMap(importedComponentCopy);
+        for (const auto &entry : newComponentNames) {
+            std::string newName = entry.first;
+            size_t count = 1;
+            while (std::find(compNames.begin(), compNames.end(), newName) != compNames.end()) {
+                newName += "_" + convertToString(count++);
+            }
+            if (newName != entry.first) {
+                entry.second->setName(newName);
+            }
+        }
+
+        // If the component 'component' has variables then they are equivalent variables and they
+        // need to be exchanged with the real variables from the component 'importedComponent'.
+        for (size_t i = 0; i < component->variableCount(); ++i) {
+            auto placeholderVariable = component->variable(i);
+            for (size_t j = 0; j < placeholderVariable->equivalentVariableCount(); ++j) {
+                auto localModelVariable = placeholderVariable->equivalentVariable(j);
+                auto importedComponentVariable = importedComponentCopy->variable(placeholderVariable->name());
+                Variable::removeEquivalence(placeholderVariable, localModelVariable);
+                Variable::addEquivalence(importedComponentVariable, localModelVariable);
+            }
+        }
+        parent->replaceComponent(index, importedComponentCopy);
+
+        // Apply the rebased equivalence map onto the modified model.
+        applyEquivalenceMapToModel(rebasedMap, model);
+
+        // Copy over units used in imported component to this model.
+        std::map<std::string, std::string> unitsNamesToReplace;
+        for (const auto &u : requiredUnits) {
+            if (!model->hasUnits(u)) {
+                auto orignalName = u->name();
+                size_t count = 0;
+                while (!model->hasUnits(u) && model->hasUnits(u->name())) {
+                    auto name = u->name();
+                    name += "_" + convertToString(++count);
+                    u->setName(name);
+                }
+                model->addUnits(u);
+                if (orignalName != u->name()) {
+                    unitsNamesToReplace[orignalName] = u->name();
+                }
+            }
+        }
+        findAndReplaceComponentsCnUnitsNames(importedComponentCopy, unitsNamesToReplace);
+    }
+}
+
+void flattenComponentTree(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t componentIndex)
+{
+    flattenComponent(parent, component, componentIndex);
+    auto flattenedComponent = parent->component(componentIndex);
+    for (size_t index = 0; index < flattenedComponent->componentCount(); ++index) {
+        auto c = flattenedComponent->component(index);
+        flattenComponentTree(flattenedComponent, c, index);
+    }
+}
+
+// Move the flatten() function here:
+ModelPtr Importer::flatten(const ModelPtr &inModel)
+{
+    if (inModel->hasUnresolvedImports()) {
+        return nullptr;
+    }
+    auto model = inModel->clone();
+
+    while (model->hasImports()) {
+        // Go through Units and instantiate any imported Units.
+        for (size_t index = 0; index < model->unitsCount(); ++index) {
+            auto u = model->units(index);
+            if (u->isImport()) {
+                auto importedUnits = u->importSource()->model()->units(u->importReference());
+                auto importedUnitsCopy = importedUnits->clone();
+                importedUnitsCopy->setName(u->name());
+                model->replaceUnits(index, importedUnitsCopy);
+            }
+        }
+
+        // Go through Components and instatiate any imported Components
+        for (size_t index = 0; index < model->componentCount(); ++index) {
+            auto c = model->component(index);
+            flattenComponentTree(model, c, index);
+        }
+    }
+
+    model->linkUnits();
+    return model;
 }
 
 } // namespace libcellml
