+++ src/model.cpp
@@ -241,96 +241,96 @@ void linkComponentVariableUnits(const ComponentPtr &component)
     }
 }
 
-NameList findCnUnitsNames(const XmlNodePtr &node)
-{
-    NameList names;
-    XmlNodePtr childNode = node->firstChild();
-    while (childNode != nullptr) {
-        if (childNode->isMathmlElement("cn")) {
-            std::string u = childNode->attribute("units");
-            if (!u.empty() && !isStandardUnitName(u)) {
-                names.push_back(u);
-            }
-        }
-        auto childNames = findCnUnitsNames(childNode);
-        names.insert(names.end(), childNames.begin(), childNames.end());
-        childNode = childNode->next();
-    }
+// NameList findCnUnitsNames(const XmlNodePtr &node)
+// {
+//     NameList names;
+//     XmlNodePtr childNode = node->firstChild();
+//     while (childNode != nullptr) {
+//         if (childNode->isMathmlElement("cn")) {
+//             std::string u = childNode->attribute("units");
+//             if (!u.empty() && !isStandardUnitName(u)) {
+//                 names.push_back(u);
+//             }
+//         }
+//         auto childNames = findCnUnitsNames(childNode);
+//         names.insert(names.end(), childNames.begin(), childNames.end());
+//         childNode = childNode->next();
+//     }
 
-    return names;
-}
+//     return names;
+// }
 
-NameList findComponentCnUnitsNames(const ComponentPtr &component)
-{
-    NameList names;
-    // Inspect the MathML in this component for any specified constant <cn> units.
-    std::string mathContent = component->math();
-    if (mathContent.empty()) {
-        return names;
-    }
-    std::vector<XmlDocPtr> mathDocs = multiRootXml(mathContent);
-    for (const auto &doc : mathDocs) {
-        auto rootNode = doc->rootNode();
-        if (rootNode->isMathmlElement("math")) {
-            auto nodesNames = findCnUnitsNames(rootNode);
-            names.insert(names.end(), nodesNames.begin(), nodesNames.end());
-        }
-    }
+// NameList findComponentCnUnitsNames(const ComponentPtr &component)
+// {
+//     NameList names;
+//     // Inspect the MathML in this component for any specified constant <cn> units.
+//     std::string mathContent = component->math();
+//     if (mathContent.empty()) {
+//         return names;
+//     }
+//     std::vector<XmlDocPtr> mathDocs = multiRootXml(mathContent);
+//     for (const auto &doc : mathDocs) {
+//         auto rootNode = doc->rootNode();
+//         if (rootNode->isMathmlElement("math")) {
+//             auto nodesNames = findCnUnitsNames(rootNode);
+//             names.insert(names.end(), nodesNames.begin(), nodesNames.end());
+//         }
+//     }
 
-    return names;
-}
+//     return names;
+// }
 
-void findAndReplaceCnUnitsNames(const XmlNodePtr &node, const StringStringMap &replaceMap)
-{
-    XmlNodePtr childNode = node->firstChild();
-    while (childNode != nullptr) {
-        if (childNode->isMathmlElement("cn")) {
-            std::string unitsName = childNode->attribute("units");
-            auto foundNameIter = replaceMap.find(unitsName);
-            if (foundNameIter != replaceMap.end()) {
-                childNode->setAttribute("units", foundNameIter->second.c_str());
-            }
-        }
-        findAndReplaceCnUnitsNames(childNode, replaceMap);
-        childNode = childNode->next();
-    }
-}
+// void findAndReplaceCnUnitsNames(const XmlNodePtr &node, const StringStringMap &replaceMap)
+// {
+//     XmlNodePtr childNode = node->firstChild();
+//     while (childNode != nullptr) {
+//         if (childNode->isMathmlElement("cn")) {
+//             std::string unitsName = childNode->attribute("units");
+//             auto foundNameIter = replaceMap.find(unitsName);
+//             if (foundNameIter != replaceMap.end()) {
+//                 childNode->setAttribute("units", foundNameIter->second.c_str());
+//             }
+//         }
+//         findAndReplaceCnUnitsNames(childNode, replaceMap);
+//         childNode = childNode->next();
+//     }
+// }
 
-void findAndReplaceComponentCnUnitsNames(const ComponentPtr &component, const StringStringMap &replaceMap)
-{
-    std::string mathContent = component->math();
-    if (mathContent.empty()) {
-        return;
-    }
-    bool contentModified = false;
-    std::string newMathContent;
-    std::vector<XmlDocPtr> mathDocs = multiRootXml(mathContent);
-    for (const auto &doc : mathDocs) {
-        auto rootNode = doc->rootNode();
-        if (rootNode->isMathmlElement("math")) {
-            auto originalMath = rootNode->convertToString();
-            findAndReplaceCnUnitsNames(rootNode, replaceMap);
-            auto newMath = rootNode->convertToString();
-            newMathContent += newMath;
-            if (newMath != originalMath) {
-                contentModified = true;
-            }
-        }
-    }
+// void findAndReplaceComponentCnUnitsNames(const ComponentPtr &component, const StringStringMap &replaceMap)
+// {
+//     std::string mathContent = component->math();
+//     if (mathContent.empty()) {
+//         return;
+//     }
+//     bool contentModified = false;
+//     std::string newMathContent;
+//     std::vector<XmlDocPtr> mathDocs = multiRootXml(mathContent);
+//     for (const auto &doc : mathDocs) {
+//         auto rootNode = doc->rootNode();
+//         if (rootNode->isMathmlElement("math")) {
+//             auto originalMath = rootNode->convertToString();
+//             findAndReplaceCnUnitsNames(rootNode, replaceMap);
+//             auto newMath = rootNode->convertToString();
+//             newMathContent += newMath;
+//             if (newMath != originalMath) {
+//                 contentModified = true;
+//             }
+//         }
+//     }
 
-    if (contentModified) {
-        component->setMath(newMathContent);
-    }
-}
+//     if (contentModified) {
+//         component->setMath(newMathContent);
+//     }
+// }
 
-void findAndReplaceComponentsCnUnitsNames(const ComponentPtr &component, const StringStringMap &replaceMap)
-{
-    findAndReplaceComponentCnUnitsNames(component, replaceMap);
-    for (size_t index = 0; index < component->componentCount(); ++index) {
-        auto childComponent = component->component(index);
-        findAndReplaceComponentCnUnitsNames(childComponent, replaceMap);
-    }
-}
+// void findAndReplaceComponentsCnUnitsNames(const ComponentPtr &component, const StringStringMap &replaceMap)
+// {
+//     findAndReplaceComponentCnUnitsNames(component, replaceMap);
+//     for (size_t index = 0; index < component->componentCount(); ++index) {
+//         auto childComponent = component->component(index);
+//         findAndReplaceComponentCnUnitsNames(childComponent, replaceMap);
+//     }
+// }
 
 void traverseComponentTreeLinkingUnits(const ComponentPtr &component)
 {
@@ -534,103 +534,103 @@ bool Model::hasImports() const
     return importsPresent;
 }
 
-size_t getComponentIndexInComponentEntity(const ComponentEntityPtr &componentParent, const ComponentEntityPtr &component)
-{
-    size_t index = 0;
-    bool found = false;
-    while (index < componentParent->componentCount() && !found) {
-        if (componentParent->component(index) == component) {
-            found = true;
-        } else {
-            ++index;
-        }
-    }
+// size_t getComponentIndexInComponentEntity(const ComponentEntityPtr &componentParent, const ComponentEntityPtr &component)
+// {
+//     size_t index = 0;
+//     bool found = false;
+//     while (index < componentParent->componentCount() && !found) {
+//         if (componentParent->component(index) == component) {
+//             found = true;
+//         } else {
+//             ++index;
+//         }
+//     }
 
-    return index;
-}
+//     return index;
+// }
 
-IndexStack reverseEngineerIndexStack(const VariablePtr &variable)
-{
-    IndexStack indexStack;
-    ComponentPtr component = std::dynamic_pointer_cast<Component>(variable->parent());
-    indexStack.push_back(getVariableIndexInComponent(component, variable));
+// IndexStack reverseEngineerIndexStack(const VariablePtr &variable)
+// {
+//     IndexStack indexStack;
+//     ComponentPtr component = std::dynamic_pointer_cast<Component>(variable->parent());
+//     indexStack.push_back(getVariableIndexInComponent(component, variable));
 
-    ComponentEntityPtr parent = component;
-    ComponentEntityPtr grandParent = std::dynamic_pointer_cast<ComponentEntity>(parent->parent());
-    while (grandParent != nullptr) {
-        indexStack.push_back(getComponentIndexInComponentEntity(grandParent, parent));
-        parent = grandParent;
-        grandParent = std::dynamic_pointer_cast<ComponentEntity>(parent->parent());
-    }
+//     ComponentEntityPtr parent = component;
+//     ComponentEntityPtr grandParent = std::dynamic_pointer_cast<ComponentEntity>(parent->parent());
+//     while (grandParent != nullptr) {
+//         indexStack.push_back(getComponentIndexInComponentEntity(grandParent, parent));
+//         parent = grandParent;
+//         grandParent = std::dynamic_pointer_cast<ComponentEntity>(parent->parent());
+//     }
 
-    std::reverse(std::begin(indexStack), std::end(indexStack));
+//     std::reverse(std::begin(indexStack), std::end(indexStack));
 
-    return indexStack;
-}
+//     return indexStack;
+// }
 
-void recordVariableEquivalences(const ComponentPtr &component, EquivalenceMap &equivalenceMap, IndexStack &indexStack)
-{
-    for (size_t index = 0; index < component->variableCount(); ++index) {
-        auto variable = component->variable(index);
-        for (size_t j = 0; j < variable->equivalentVariableCount(); ++j) {
-            if (j == 0) {
-                indexStack.push_back(index);
-            }
-            auto equivalentVariable = variable->equivalentVariable(j);
-            auto equivalentVariableIndexStack = reverseEngineerIndexStack(equivalentVariable);
-            if (equivalenceMap.count(indexStack) == 0) {
-                equivalenceMap[indexStack] = std::vector<IndexStack>();
-            }
-            equivalenceMap[indexStack].push_back(equivalentVariableIndexStack);
-        }
-        if (variable->equivalentVariableCount() > 0) {
-            indexStack.pop_back();
-        }
-    }
-}
+// void recordVariableEquivalences(const ComponentPtr &component, EquivalenceMap &equivalenceMap, IndexStack &indexStack)
+// {
+//     for (size_t index = 0; index < component->variableCount(); ++index) {
+//         auto variable = component->variable(index);
+//         for (size_t j = 0; j < variable->equivalentVariableCount(); ++j) {
+//             if (j == 0) {
+//                 indexStack.push_back(index);
+//             }
+//             auto equivalentVariable = variable->equivalentVariable(j);
+//             auto equivalentVariableIndexStack = reverseEngineerIndexStack(equivalentVariable);
+//             if (equivalenceMap.count(indexStack) == 0) {
+//                 equivalenceMap[indexStack] = std::vector<IndexStack>();
+//             }
+//             equivalenceMap[indexStack].push_back(equivalentVariableIndexStack);
+//         }
+//         if (variable->equivalentVariableCount() > 0) {
+//             indexStack.pop_back();
+//         }
+//     }
+// }
 
-void generateEquivalenceMap(const ComponentPtr &component, EquivalenceMap &map, IndexStack &indexStack)
-{
-    for (size_t index = 0; index < component->componentCount(); ++index) {
-        indexStack.push_back(index);
-        auto c = component->component(index);
-        recordVariableEquivalences(c, map, indexStack);
-        generateEquivalenceMap(c, map, indexStack);
-        indexStack.pop_back();
-    }
-}
+// void generateEquivalenceMap(const ComponentPtr &component, EquivalenceMap &map, IndexStack &indexStack)
+// {
+//     for (size_t index = 0; index < component->componentCount(); ++index) {
+//         indexStack.push_back(index);
+//         auto c = component->component(index);
+//         recordVariableEquivalences(c, map, indexStack);
+//         generateEquivalenceMap(c, map, indexStack);
+//         indexStack.pop_back();
+//     }
+// }
 
-VariablePtr getVariableLocatedAt(const IndexStack &stack, const ModelPtr &model)
-{
-    ComponentPtr component;
-    for (size_t index = 0; index < stack.size() - 1; ++index) {
-        if (index == 0) {
-            component = model->component(stack.at(index));
-        } else {
-            component = component->component(stack.at(index));
-        }
-    }
+// VariablePtr getVariableLocatedAt(const IndexStack &stack, const ModelPtr &model)
+// {
+//     ComponentPtr component;
+//     for (size_t index = 0; index < stack.size() - 1; ++index) {
+//         if (index == 0) {
+//             component = model->component(stack.at(index));
+//         } else {
+//             component = component->component(stack.at(index));
+//         }
+//     }
 
-    return component->variable(stack.back());
-}
+//     return component->variable(stack.back());
+// }
 
-void makeEquivalence(const IndexStack &stack1, const IndexStack &stack2, const ModelPtr &model)
-{
-    auto v1 = getVariableLocatedAt(stack1, model);
-    auto v2 = getVariableLocatedAt(stack2, model);
-    Variable::addEquivalence(v1, v2);
-}
+// void makeEquivalence(const IndexStack &stack1, const IndexStack &stack2, const ModelPtr &model)
+// {
+//     auto v1 = getVariableLocatedAt(stack1, model);
+//     auto v2 = getVariableLocatedAt(stack2, model);
+//     Variable::addEquivalence(v1, v2);
+// }
 
-void applyEquivalenceMapToModel(const EquivalenceMap &map, const ModelPtr &model)
-{
-    for (const auto &iter : map) {
-        auto key = iter.first;
-        auto vector = iter.second;
-        for (auto vectorIter = vector.begin(); vectorIter < vector.end(); ++vectorIter) {
-            makeEquivalence(key, *vectorIter, model);
-        }
-    }
-}
+// void applyEquivalenceMapToModel(const EquivalenceMap &map, const ModelPtr &model)
+// {
+//     for (const auto &iter : map) {
+//         auto key = iter.first;
+//         auto vector = iter.second;
+//         for (auto vectorIter = vector.begin(); vectorIter < vector.end(); ++vectorIter) {
+//             makeEquivalence(key, *vectorIter, model);
+//         }
+//     }
+// }
 
 ModelPtr Model::clone() const
 {
@@ -664,275 +664,275 @@ ModelPtr Model::clone() const
     return m;
 }
 
-IndexStack reverseEngineerIndexStack(const ComponentPtr &component)
-{
-    auto dummyVariable = Variable::create();
-    component->addVariable(dummyVariable);
-    IndexStack indexStack = reverseEngineerIndexStack(dummyVariable);
-    indexStack.pop_back();
-    component->removeVariable(dummyVariable);
+// IndexStack reverseEngineerIndexStack(const ComponentPtr &component)
+// {
+//     auto dummyVariable = Variable::create();
+//     component->addVariable(dummyVariable);
+//     IndexStack indexStack = reverseEngineerIndexStack(dummyVariable);
+//     indexStack.pop_back();
+//     component->removeVariable(dummyVariable);
 
-    return indexStack;
-}
+//     return indexStack;
+// }
 
-IndexStack rebaseIndexStack(const IndexStack &stack, const IndexStack &originStack, const IndexStack &destinationStack)
-{
-    auto rebasedStack = stack;
+// IndexStack rebaseIndexStack(const IndexStack &stack, const IndexStack &originStack, const IndexStack &destinationStack)
+// {
+//     auto rebasedStack = stack;
 
-    rebasedStack.resize(originStack.size(), SIZE_MAX);
-    if (rebasedStack == originStack) {
-        rebasedStack = destinationStack;
-        auto offsetIt = stack.begin() + static_cast<int64_t>(originStack.size());
-        rebasedStack.insert(rebasedStack.end(), offsetIt, stack.end());
-    } else {
-        rebasedStack.clear();
-    }
+//     rebasedStack.resize(originStack.size(), SIZE_MAX);
+//     if (rebasedStack == originStack) {
+//         rebasedStack = destinationStack;
+//         auto offsetIt = stack.begin() + static_cast<int64_t>(originStack.size());
+//         rebasedStack.insert(rebasedStack.end(), offsetIt, stack.end());
+//     } else {
+//         rebasedStack.clear();
+//     }
 
-    return rebasedStack;
-}
+//     return rebasedStack;
+// }
 
-EquivalenceMap rebaseEquivalenceMap(const EquivalenceMap &map, const IndexStack &originStack, const IndexStack &destinationStack)
-{
-    EquivalenceMap rebasedMap;
-    for (const auto &entry : map) {
-        auto key = entry.first;
-        auto rebasedKey = rebaseIndexStack(key, originStack, destinationStack);
-        if (!rebasedKey.empty()) {
-            auto vector = entry.second;
-            std::vector<IndexStack> rebasedVector;
-            for (const auto &stack : vector) {
-                auto rebasedTarget = rebaseIndexStack(stack, originStack, destinationStack);
-                if (!rebasedTarget.empty()) {
-                    rebasedVector.push_back(rebasedTarget);
-                }
-            }
+// EquivalenceMap rebaseEquivalenceMap(const EquivalenceMap &map, const IndexStack &originStack, const IndexStack &destinationStack)
+// {
+//     EquivalenceMap rebasedMap;
+//     for (const auto &entry : map) {
+//         auto key = entry.first;
+//         auto rebasedKey = rebaseIndexStack(key, originStack, destinationStack);
+//         if (!rebasedKey.empty()) {
+//             auto vector = entry.second;
+//             std::vector<IndexStack> rebasedVector;
+//             for (const auto &stack : vector) {
+//                 auto rebasedTarget = rebaseIndexStack(stack, originStack, destinationStack);
+//                 if (!rebasedTarget.empty()) {
+//                     rebasedVector.push_back(rebasedTarget);
+//                 }
+//             }
 
-            if (!rebasedVector.empty()) {
-                rebasedMap[rebasedKey] = rebasedVector;
-            }
-        }
-    }
+//             if (!rebasedVector.empty()) {
+//                 rebasedMap[rebasedKey] = rebasedVector;
+//             }
+//         }
+//     }
 
-    return rebasedMap;
-}
+//     return rebasedMap;
+// }
 
-void componentNames(const ComponentPtr &component, NameList &names)
-{
-    for (size_t index = 0; index < component->componentCount(); ++index) {
-        auto c = component->component(index);
-        names.push_back(c->name());
-        componentNames(c, names);
-    }
-}
+// void componentNames(const ComponentPtr &component, NameList &names)
+// {
+//     for (size_t index = 0; index < component->componentCount(); ++index) {
+//         auto c = component->component(index);
+//         names.push_back(c->name());
+//         componentNames(c, names);
+//     }
+// }
 
-NameList componentNames(const ModelPtr &model)
-{
-    NameList names;
-    for (size_t index = 0; index < model->componentCount(); ++index) {
-        auto component = model->component(index);
-        names.push_back(component->name());
-        componentNames(component, names);
-    }
-    return names;
-}
+// NameList componentNames(const ModelPtr &model)
+// {
+//     NameList names;
+//     for (size_t index = 0; index < model->componentCount(); ++index) {
+//         auto component = model->component(index);
+//         names.push_back(component->name());
+//         componentNames(component, names);
+//     }
+//     return names;
+// }
 
-ComponentNameMap createComponentNamesMap(const ComponentPtr &component)
-{
-    ComponentNameMap nameMap;
-    for (size_t index = 0; index < component->componentCount(); ++index) {
-        auto c = component->component(index);
-        nameMap[c->name()] = c;
-        ComponentNameMap childrenNameMap = createComponentNamesMap(c);
-        nameMap.insert(childrenNameMap.begin(), childrenNameMap.end());
-    }
+// ComponentNameMap createComponentNamesMap(const ComponentPtr &component)
+// {
+//     ComponentNameMap nameMap;
+//     for (size_t index = 0; index < component->componentCount(); ++index) {
+//         auto c = component->component(index);
+//         nameMap[c->name()] = c;
+//         ComponentNameMap childrenNameMap = createComponentNamesMap(c);
+//         nameMap.insert(childrenNameMap.begin(), childrenNameMap.end());
+//     }
 
-    return nameMap;
-}
+//     return nameMap;
+// }
 
-std::vector<UnitsPtr> referencedUnits(const ModelPtr &model, const UnitsPtr &units)
-{
-    std::vector<UnitsPtr> requiredUnits;
+// std::vector<UnitsPtr> referencedUnits(const ModelPtr &model, const UnitsPtr &units)
+// {
+//     std::vector<UnitsPtr> requiredUnits;
 
-    std::string ref;
-    std::string pre;
-    std::string id;
-    double expMult;
-    double uExp;
+//     std::string ref;
+//     std::string pre;
+//     std::string id;
+//     double expMult;
+//     double uExp;
 
-    for (size_t index = 0; index < units->unitCount(); ++index) {
-        units->unitAttributes(index, ref, pre, uExp, expMult, id);
-        if (!isStandardUnitName(ref)) {
-            auto refUnits = model->units(ref);
-            if (refUnits != nullptr) {
-                auto requiredUnitsUnits = referencedUnits(model, refUnits);
-                requiredUnits.insert(requiredUnits.end(), requiredUnitsUnits.begin(), requiredUnitsUnits.end());
-                requiredUnits.push_back(refUnits);
-            }
-        }
-    }
+//     for (size_t index = 0; index < units->unitCount(); ++index) {
+//         units->unitAttributes(index, ref, pre, uExp, expMult, id);
+//         if (!isStandardUnitName(ref)) {
+//             auto refUnits = model->units(ref);
+//             if (refUnits != nullptr) {
+//                 auto requiredUnitsUnits = referencedUnits(model, refUnits);
+//                 requiredUnits.insert(requiredUnits.end(), requiredUnitsUnits.begin(), requiredUnitsUnits.end());
+//                 requiredUnits.push_back(refUnits);
+//             }
+//         }
+//     }
 
-    return requiredUnits;
-}
+//     return requiredUnits;
+// }
 
-std::vector<UnitsPtr> unitsUsed(const ModelPtr &model, const ComponentPtr &component)
-{
-    std::vector<UnitsPtr> usedUnits;
-    for (size_t i = 0; i < component->variableCount(); ++i) {
-        auto v = component->variable(i);
-        auto u = v->units();
-        if (u != nullptr && !isStandardUnitName(u->name())) {
-            auto requiredUnits = referencedUnits(model, u);
-            usedUnits.insert(usedUnits.end(), requiredUnits.begin(), requiredUnits.end());
-            usedUnits.push_back(u);
-        }
-    }
-    auto componentCnUnitsNames = findComponentCnUnitsNames(component);
-    for (const auto &unitsName : componentCnUnitsNames) {
-        auto u = model->units(unitsName);
-        if (u != nullptr && !isStandardUnitName(u->name())) {
-            auto requiredUnits = referencedUnits(model, u);
-            usedUnits.insert(usedUnits.end(), requiredUnits.begin(), requiredUnits.end());
-            usedUnits.push_back(u);
-        }
-    }
+// std::vector<UnitsPtr> unitsUsed(const ModelPtr &model, const ComponentPtr &component)
+// {
+//     std::vector<UnitsPtr> usedUnits;
+//     for (size_t i = 0; i < component->variableCount(); ++i) {
+//         auto v = component->variable(i);
+//         auto u = v->units();
+//         if (u != nullptr && !isStandardUnitName(u->name())) {
+//             auto requiredUnits = referencedUnits(model, u);
+//             usedUnits.insert(usedUnits.end(), requiredUnits.begin(), requiredUnits.end());
+//             usedUnits.push_back(u);
+//         }
+//     }
+//     auto componentCnUnitsNames = findComponentCnUnitsNames(component);
+//     for (const auto &unitsName : componentCnUnitsNames) {
+//         auto u = model->units(unitsName);
+//         if (u != nullptr && !isStandardUnitName(u->name())) {
+//             auto requiredUnits = referencedUnits(model, u);
+//             usedUnits.insert(usedUnits.end(), requiredUnits.begin(), requiredUnits.end());
+//             usedUnits.push_back(u);
+//         }
+//     }
 
-    for (size_t i = 0; i < component->componentCount(); ++i) {
-        auto childComponent = component->component(i);
-        auto childUsedUnits = unitsUsed(model, childComponent);
-        usedUnits.insert(usedUnits.end(), childUsedUnits.begin(), childUsedUnits.end());
-    }
+//     for (size_t i = 0; i < component->componentCount(); ++i) {
+//         auto childComponent = component->component(i);
+//         auto childUsedUnits = unitsUsed(model, childComponent);
+//         usedUnits.insert(usedUnits.end(), childUsedUnits.begin(), childUsedUnits.end());
+//     }
 
-    return usedUnits;
-}
+//     return usedUnits;
+// }
 
-void flattenComponent(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t index)
-{
-    if (component->isImport()) {
-        auto model = owningModel(component);
-        auto importSource = component->importSource();
-        auto importModel = importSource->model();
-        auto importedComponent = importModel->component(component->importReference());
+// void flattenComponent(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t index)
+// {
+//     if (component->isImport()) {
+//         auto model = owningModel(component);
+//         auto importSource = component->importSource();
+//         auto importModel = importSource->model();
+//         auto importedComponent = importModel->component(component->importReference());
 
-        // Determine names of components already in use.
-        NameList compNames = componentNames(model);
+//         // Determine names of components already in use.
+//         NameList compNames = componentNames(model);
 
-        // Determine the stack for the destination component.
-        IndexStack destinationComponentBaseIndexStack = reverseEngineerIndexStack(component);
+//         // Determine the stack for the destination component.
+//         IndexStack destinationComponentBaseIndexStack = reverseEngineerIndexStack(component);
 
-        // Determine the stack for the source component.
-        IndexStack importedComponentBaseIndexStack = reverseEngineerIndexStack(importedComponent);
+//         // Determine the stack for the source component.
+//         IndexStack importedComponentBaseIndexStack = reverseEngineerIndexStack(importedComponent);
 
-        // Generate equivalence map for the source component.
-        EquivalenceMap map;
-        recordVariableEquivalences(importedComponent, map, importedComponentBaseIndexStack);
-        generateEquivalenceMap(importedComponent, map, importedComponentBaseIndexStack);
+//         // Generate equivalence map for the source component.
+//         EquivalenceMap map;
+//         recordVariableEquivalences(importedComponent, map, importedComponentBaseIndexStack);
+//         generateEquivalenceMap(importedComponent, map, importedComponentBaseIndexStack);
 
-        // Rebase the generated equivalence map from the source component to the destination component.
-        auto rebasedMap = rebaseEquivalenceMap(map, importedComponentBaseIndexStack, destinationComponentBaseIndexStack);
+//         // Rebase the generated equivalence map from the source component to the destination component.
+//         auto rebasedMap = rebaseEquivalenceMap(map, importedComponentBaseIndexStack, destinationComponentBaseIndexStack);
 
-        // Take a copy of the imported component which will be used to replace the import defined in this model.
-        auto importedComponentCopy = importedComponent->clone();
-        importedComponentCopy->setName(component->name());
-        for (size_t i = 0; i < component->componentCount(); ++i) {
-            importedComponentCopy->addComponent(component->component(i));
-        }
+//         // Take a copy of the imported component which will be used to replace the import defined in this model.
+//         auto importedComponentCopy = importedComponent->clone();
+//         importedComponentCopy->setName(component->name());
+//         for (size_t i = 0; i < component->componentCount(); ++i) {
+//             importedComponentCopy->addComponent(component->component(i));
+//         }
 
-        // Get list of required units from component's variables.
-        std::vector<UnitsPtr> requiredUnits = unitsUsed(importModel, importedComponentCopy);
+//         // Get list of required units from component's variables.
+//         std::vector<UnitsPtr> requiredUnits = unitsUsed(importModel, importedComponentCopy);
 
-        // Add all required units to a model so referenced units can be resolved.
-        auto requiredUnitsModel = Model::create();
-        for (const auto &units : requiredUnits) {
-            requiredUnitsModel->addUnits(units);
-        }
+//         // Add all required units to a model so referenced units can be resolved.
+//         auto requiredUnitsModel = Model::create();
+//         for (const auto &units : requiredUnits) {
+//             requiredUnitsModel->addUnits(units);
+//         }
 
-        // Make a map of component name to component pointer.
-        ComponentNameMap newComponentNames = createComponentNamesMap(importedComponentCopy);
-        for (const auto &entry : newComponentNames) {
-            std::string newName = entry.first;
-            size_t count = 1;
-            while (std::find(compNames.begin(), compNames.end(), newName) != compNames.end()) {
-                newName += "_" + convertToString(count++);
-            }
-            if (newName != entry.first) {
-                entry.second->setName(newName);
-            }
-        }
+//         // Make a map of component name to component pointer.
+//         ComponentNameMap newComponentNames = createComponentNamesMap(importedComponentCopy);
+//         for (const auto &entry : newComponentNames) {
+//             std::string newName = entry.first;
+//             size_t count = 1;
+//             while (std::find(compNames.begin(), compNames.end(), newName) != compNames.end()) {
+//                 newName += "_" + convertToString(count++);
+//             }
+//             if (newName != entry.first) {
+//                 entry.second->setName(newName);
+//             }
+//         }
 
-        // If the component 'component' has variables then they are equivalent variables and they
-        // need to be exchanged with the real variables from the component 'importedComponent'.
-        for (size_t i = 0; i < component->variableCount(); ++i) {
-            auto placeholderVariable = component->variable(i);
-            for (size_t j = 0; j < placeholderVariable->equivalentVariableCount(); ++j) {
-                auto localModelVariable = placeholderVariable->equivalentVariable(j);
-                auto importedComponentVariable = importedComponentCopy->variable(placeholderVariable->name());
-                Variable::removeEquivalence(placeholderVariable, localModelVariable);
-                Variable::addEquivalence(importedComponentVariable, localModelVariable);
-            }
-        }
-        parent->replaceComponent(index, importedComponentCopy);
+//         // If the component 'component' has variables then they are equivalent variables and they
+//         // need to be exchanged with the real variables from the component 'importedComponent'.
+//         for (size_t i = 0; i < component->variableCount(); ++i) {
+//             auto placeholderVariable = component->variable(i);
+//             for (size_t j = 0; j < placeholderVariable->equivalentVariableCount(); ++j) {
+//                 auto localModelVariable = placeholderVariable->equivalentVariable(j);
+//                 auto importedComponentVariable = importedComponentCopy->variable(placeholderVariable->name());
+//                 Variable::removeEquivalence(placeholderVariable, localModelVariable);
+//                 Variable::addEquivalence(importedComponentVariable, localModelVariable);
+//             }
+//         }
+//         parent->replaceComponent(index, importedComponentCopy);
 
-        // Apply the rebased equivalence map onto the modified model.
-        applyEquivalenceMapToModel(rebasedMap, model);
+//         // Apply the rebased equivalence map onto the modified model.
+//         applyEquivalenceMapToModel(rebasedMap, model);
 
-        // Copy over units used in imported component to this model.
-        std::map<std::string, std::string> unitsNamesToReplace;
-        for (const auto &u : requiredUnits) {
-            if (!model->hasUnits(u)) {
-                auto orignalName = u->name();
-                size_t count = 0;
-                while (!model->hasUnits(u) && model->hasUnits(u->name())) {
-                    auto name = u->name();
-                    name += "_" + convertToString(++count);
-                    u->setName(name);
-                }
-                model->addUnits(u);
-                if (orignalName != u->name()) {
-                    unitsNamesToReplace[orignalName] = u->name();
-                }
-            }
-        }
-        findAndReplaceComponentsCnUnitsNames(importedComponentCopy, unitsNamesToReplace);
-    }
-}
+//         // Copy over units used in imported component to this model.
+//         std::map<std::string, std::string> unitsNamesToReplace;
+//         for (const auto &u : requiredUnits) {
+//             if (!model->hasUnits(u)) {
+//                 auto orignalName = u->name();
+//                 size_t count = 0;
+//                 while (!model->hasUnits(u) && model->hasUnits(u->name())) {
+//                     auto name = u->name();
+//                     name += "_" + convertToString(++count);
+//                     u->setName(name);
+//                 }
+//                 model->addUnits(u);
+//                 if (orignalName != u->name()) {
+//                     unitsNamesToReplace[orignalName] = u->name();
+//                 }
+//             }
+//         }
+//         findAndReplaceComponentsCnUnitsNames(importedComponentCopy, unitsNamesToReplace);
+//     }
+// }
 
-void flattenComponentTree(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t componentIndex)
-{
-    flattenComponent(parent, component, componentIndex);
-    auto flattenedComponent = parent->component(componentIndex);
-    for (size_t index = 0; index < flattenedComponent->componentCount(); ++index) {
-        auto c = flattenedComponent->component(index);
-        flattenComponentTree(flattenedComponent, c, index);
-    }
-}
+// void flattenComponentTree(const ComponentEntityPtr &parent, const ComponentPtr &component, size_t componentIndex)
+// {
+//     flattenComponent(parent, component, componentIndex);
+//     auto flattenedComponent = parent->component(componentIndex);
+//     for (size_t index = 0; index < flattenedComponent->componentCount(); ++index) {
+//         auto c = flattenedComponent->component(index);
+//         flattenComponentTree(flattenedComponent, c, index);
+//     }
+// }
 
-void Model::flatten()
-{
-    if (hasUnresolvedImports()) {
-        return;
-    }
+// void Model::flatten()
+// {
+//     if (hasUnresolvedImports()) {
+//         return;
+//     }
 
-    while (hasImports()) {
-        // Go through Units and instantiate any imported Units.
-        for (size_t index = 0; index < unitsCount(); ++index) {
-            auto u = units(index);
-            if (u->isImport()) {
-                auto importedUnits = u->importSource()->model()->units(u->importReference());
-                auto importedUnitsCopy = importedUnits->clone();
-                importedUnitsCopy->setName(u->name());
-                replaceUnits(index, importedUnitsCopy);
-            }
-        }
+//     while (hasImports()) {
+//         // Go through Units and instantiate any imported Units.
+//         for (size_t index = 0; index < unitsCount(); ++index) {
+//             auto u = units(index);
+//             if (u->isImport()) {
+//                 auto importedUnits = u->importSource()->model()->units(u->importReference());
+//                 auto importedUnitsCopy = importedUnits->clone();
+//                 importedUnitsCopy->setName(u->name());
+//                 replaceUnits(index, importedUnitsCopy);
+//             }
+//         }
 
-        // Go through Components and instatiate any imported Components
-        for (size_t index = 0; index < componentCount(); ++index) {
-            auto c = component(index);
-            flattenComponentTree(shared_from_this(), c, index);
-        }
-    }
+//         // Go through Components and instatiate any imported Components
+//         for (size_t index = 0; index < componentCount(); ++index) {
+//             auto c = component(index);
+//             flattenComponentTree(shared_from_this(), c, index);
+//         }
+//     }
 
-    linkUnits();
-}
+//     linkUnits();
+// }
 
 bool Model::fixVariableInterfaces()
 {
